+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 ### Added
+- Shortcut for postman generation on a script (<kbd>Alt</kbd>+<kbd>P</kbd>)
+- Action for directly deploy a script to an environment (context menu or <kbd>Alt</kbd>+<kbd>D</kbd>)
 
 ### Changed